test(backup/gcs): split manifest tests by concern
